move iceberg table information into new file
