+++ src/storage/irc_table_entry.cpp
@@ -17,7 +17,7 @@
 #include "duckdb/planner/tableref/bound_at_clause.hpp"
 
 #include "rest_catalog/objects/list.hpp"
-#include "storage/irc_transaction.hpp"
+#include "storage/iceberg_table_information.hpp"
 
 namespace duckdb {
